// src/graph/types.ts
export type Maybe<T> = T | null;

export interface SearchIntent {
  id: Maybe<number>;

  agentId: Maybe<number>;

  isTopLevel: Maybe<boolean>;
}

export interface SearchIntentTrigger {
  id: Maybe<number>;

  value: Maybe<string>;

  type: Maybe<string>;
}

export interface IntentExpression {
  intentExpression: Maybe<string>;

  agentId: Maybe<number>;

  isTopLevel: Maybe<boolean>;
}

export interface AgentId {
  agentId: number;
}

export interface SearchSingleDictionary {
  id: number;
}

export interface CreateIntent {
  agentId: number;

  name: string;

  parentId: Maybe<number>;

  triggers: CreateIntentTriggerDataInput[];

  outputs: Json[];
}

export interface CreateIntentTriggerDataInput {
  type: IntentTriggerType;

  value: string[];
}

export interface UpdateIntent {
  id: number;

  name: Maybe<string>;

  isTopLevel: Maybe<boolean>;

  parentId: Maybe<number>;

  triggers: Maybe<IntentTriggerDataInput[]>;

  outputs: Maybe<Json[]>;
}

export interface IntentTriggerDataInput {
  id: Maybe<number>;

  type: Maybe<IntentTriggerType>;

  value: Maybe<string[]>;

  actionType: ActionType;
}
// ...
export enum IntentTriggerType {
  Plain = "PLAIN",
  Combination = "COMBINATION",
}

export enum ActionType {
  Create = "CREATE",
  Update = "UPDATE",
  Delete = "DELETE",
}